+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/attributes/url_attributes.py
@@ -32,7 +32,7 @@ In such case username and password SHOULD be redacted and attribute's value SHOU
 
 Sensitive content provided in `url.full` SHOULD be scrubbed when instrumentations can identify it.
 
-![Experimental](https://img.shields.io/badge/-experimental-blue)
+![Development](https://img.shields.io/badge/-development-blue)
 Query string values for the following keys SHOULD be redacted by default and replaced by the
 value `REDACTED`:
 
@@ -58,7 +58,7 @@ URL_QUERY: Final = "url.query"
 The [URI query](https://www.rfc-editor.org/rfc/rfc3986#section-3.4) component.
 Note: Sensitive content provided in `url.query` SHOULD be scrubbed when instrumentations can identify it.
 
-![Experimental](https://img.shields.io/badge/-experimental-blue)
+![Development](https://img.shields.io/badge/-development-blue)
 Query string values for the following keys SHOULD be redacted by default and replaced by the value `REDACTED`:
 
 * [`AWSAccessKeyId`](https://docs.aws.amazon.com/AmazonS3/latest/userguide/RESTAuthentication.html#RESTAuthenticationQueryStringAuth)