+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Fix user agent in OTLP HTTP metrics exporter
+  ([#4475](https://github.com/open-telemetry/opentelemetry-python/pull/4475))
+
+## Version 1.31.0/0.52b0 (2025-03-12)
+
+- semantic-conventions: Bump to 1.31.0
+  ([#4471](https://github.com/open-telemetry/opentelemetry-python/pull/4471))
 - Add type annotations to context's attach & detach
   ([#4346](https://github.com/open-telemetry/opentelemetry-python/pull/4346))
 - Fix OTLP encoders missing instrumentation scope schema url and attributes
@@ -32,6 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#4406](https://github.com/open-telemetry/opentelemetry-python/pull/4406))
 - Fix env var error message for TraceLimits/SpanLimits
   ([#4458](https://github.com/open-telemetry/opentelemetry-python/pull/4458))
+- pylint-ci updated python version to 3.13
+  ([#4450](https://github.com/open-telemetry/opentelemetry-python/pull/4450))
 
 ## Version 1.30.0/0.51b0 (2025-02-03)